extract `URLProtocolStub` to separate file
